Moved all components into components directory. Additionally created custom question object with mongoose and refactored create project function to work with the custom question object
The Angular components with all in the app folder, so I moved them into a components directory for better organization. Minor changes to the post creation page, but this commit is necessary to keep everyone on the same page.

Created a custom mongoose object for the databaseto handle questions created by the faculty. It contains will allow for the frontend to recreate the information/forms while only using one object. Additionally refactored the create project function to work with the new object.